backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
supabase==2.5.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.2
apscheduler==3.10.4
tweepy==4.14.0
nba_api==1.4.1
pytest==8.2.2
pytest-asyncio==0.23.7
